Sidenav bar updated to common component
