+++ mock-relying-party-ui/src/pages/UserProfile.js
@@ -1,11 +1,12 @@
 import React from "react";
 import relyingPartyService from "../services/relyingPartyService";
 import ProfileUI from "../components/ProfileUI";
-
+import Sidenav from "../components/Sidenav";
 
 export default function UserProfilePage({ langOptions }) {
   return (
-    <ProfileUI
-      relyingPartyService={relyingPartyService} />
+    <Sidenav relyingPartyService={relyingPartyService}  component={React.createElement(ProfileUI, {
+      relyingPartyService:relyingPartyService
+    })}/>
   );
 }
